+++ src/main.rs
@@ -1,3 +1,6 @@
+mod lexer;
+
 fn main() {
     println!("Hello, world!");
+    lexer::tokenize("123 455");
 }